Add content deletion request handling workflow to "Moderator Instructions"
Previously, the forum assets provided post templates for handling content deletion requests, but did not provide
instructions for handling the requests.

A comprehensive workflow is now provided in the "Moderator Instructions" asset topic, as well as templates for each post
specified by the workflow. These supersede the previous post templates and so those templates are removed.