Add "Conserve the Conversation" section to FAQ/Guidelines asset topic
A common problem on the forum is newcomers causing confusion and frustration by editing or deleting their posts after
others have replied to them.

This is typically not done out of any malicious intentions so it will be useful to explain that it is not appropriate,
and why, in the forum Guidelines.

One of the common reasons for users later deleting a post is that they wish to reduce their "digital footprint". In this
case, the correct solution will be to request the anonymization of the post rather than deleting it. So a dedicated
subsection is added to address this subject.