Add post asset source comments to "Moderator Instructions" post
The "Moderator Instructions" post is illustrated with several graphics and diagrams.

The image files of the published post are hosted on the forum server, and so it is those URLs that are referenced in the
image embedding markup. However, the source files for those assets are hosted in the arduino/forum-assets GitHub
repository, and it is these files that should be worked with for any updates of the post.

Previously there was no indication in the post content of the source of its assets, and thus a contributor might not be
aware or their existence, or where to find them. This is resolved by adding adjacent comments to the post content with
links to the location of each of the asset source files.