Control Python version used by "Check Category Order Metadata" workflow
Previously, the workflow would use whichever version of Python was installed on the GitHub Actions runner machine. This
meant that an update to the runner environment could break the workflow at any time.

The "actions/setup-python" action is now used to install the project's standardized version of Python, as specified by
the pyproject.toml metadata file.