Rename "Avrdude, stk500, Bootloader issues" to "Uploading"
The previous category name was chosen at a time when only AVR boards were used in the Arduino ecosystem, without the
foresight to consider that it was not reasonable to define such a narrow scope for the category.

The new name better reflects the actual scope of the category and avoids excessive verbosity.