Rename "Local Groups" category to "Groups and Events"
Previously, there was a separate category for discussions related to groups (unnecessarily limited to "local"), and another "Workshops and Events" for discussions related to events. It is useful to have a dedicated place for such discussions, but the usage of both categories is extremely low. Having a large number of categories is harmful to the user experience, so we are merging categories where reasonable. Since these two categories are somewhat related, it will make more sense to have a single category for these subject matters.

The first step towards that merge is to expand the scope of one of the two categories to cover both. The redundant "Workshops and Events" category has already been closed to new topics

The next steps will be:

1. Move all topics from "Workshops and Events" to "Groups and Events"
1. Set up a redirect from the URL of the "Workshops and Events" category to the "Groups and Events" category.
1. Delete the "Workshops and Events" category.