Deprecate "Workshops and Events" category
The category has now been merged with the "Groups and Events" category.

This is an interim measure until we can set up a redirect from the "Workshops and Events" category URL to the "Groups
and Events" category in order to avoid breaking any links. Once that is done, the "Workshops and Events" category will
be deleted.