Deprecate "India" category
Due to lack of use and need (since English language was used in nearly all posts in the category), the category is being
removed. All posts have been moved to the appropriate subject-specific categories.

This is an interim measure until we can set up a redirect from the "India" category URL to the categories home page in
order to avoid breaking any links. Once that is done, the "India" category will be deleted.