Remove "Policies" category
This internal category has been deprecated for years, and was rarely used even prior to the deprecation, with much of
the use not being on the subject of policies. So it is hereby deleted.

Since it is a private category, there is not the concern about breaking links as is the case with public categories, so
we don't need to put this on hold until a redirect can be set up.