Deprecate "Emergency Response: Remote Learning" category
Due to lack of use and need (since we already have an "Education and Teaching" category), the category is being removed.
All posts have been moved to the appropriate subject-specific categories.

This is an interim measure until we can set up a redirect from the "Emergency Response: Remote Learning" category URL to
the "Education and Teaching" category in order to avoid breaking any links. Once that is done, the "Emergency Response:
Remote Learning" category will be deleted.