Deprecate "Projects Discussion and Showcase" category
This category is being removed as superfluous, after its child categories have either been moved to more appropriate parent categories, or merged away.

The category was purely a container, it no longer contains any posts.

This is an interim measure until we can set up a redirect from the "Projects Discussion and Showcase" category URL to
the "Projects" category in order to avoid breaking any links. Once that is done, the "Projects Discussion and Showcase"
category will be deleted.