Rename "Project Guidance" to "General Guidance"
The category has been moved under the "Projects" parent category, which rendered the "project" in the previous name
redundant. Replacing this with "general" better communicates the purpose of the category as a place for discussions that
involve a project as a whole, where it will not fit within the scope of the more specific forum categories.