Deprecate "Development" category
This category is being removed as superfluous, after its child categories have either been moved to more appropriate parent categories, or merged away.

The category was purely a container, it no longer contains any posts.

This is an interim measure until we can set up a redirect from the "Development" category URL to
the categories home page in order to avoid breaking any links. Once that is done, the "Development" category will be
deleted.