Restore "Uncategorized" category to first position in category listing
The "Uncategorized" category is used to catch the topics created by users who arbitrarily select the first category on
the menu (either due to laziness, or being a spammer).

The category's position was changed by the adjustment to the relative order of its parent container category. It is
hereby moved to the new first top level container category to allow it to continue to serve its function.