Reduce verbosity of "Other Hardware Development" category name
The "Other" in the "Other Hardware Development" category name was always superfluous, since it is likely implicit and
anyway there was no reason to limit the scope to "other hardware" anyway.

The superfluity of this word was increased greatly when the category was moved to the new "Other Hardware" parent
category.

Unnecessary verbosity of category names is harmful because it increases the chance that the category name will be
truncated in category listings in the forum UI.

So the category is hereby renamed to "Hardware Development". This (especially since it is generally shown in the context
of the parent category) will be entirely sufficient to communicate its purpose.