Update name of "Portenta" product class category to match official naming
The subcategories of the "Official Hardware" category are organized under a subcategory for each of the product classes.
These product classes are defined by the hardware documentation home page.

That hardware documentation uses the name "Portenta Family" for the class, so the forum's category name is updated to
match. This will also more effectively communicate the nature of the category as a parent for specific product
categories.