Update category slugs in URLs
Some of the forum categories were renamed. The subcategories have the slug of the parent category in their URL. Some of
these slugs were not updated at the time of the category renaming.

The slug component of the URL is completely ignored by the forum software, with the ID component being the only
technically relevant part. So the link check did not catch these missed updates.