Add "Classic" product class category
The subcategories of the "Official Hardware" category are to be organized under a subcategory for each of the product
classes. These product classes are defined by the hardware documentation home page.

That hardware documentation places the foundational Arduino boards under a class named "Classic", so a forum category of
this name is added