Move "Mega" class product categories to appropriate parent category
The subcategories of the "Official Hardware" category are to be organized under a subcategory for each of the product
classes. These product classes are defined by the hardware documentation home page.

That hardware documentation places the products that use the "Mega" form factor under a class named "Mega".

Previously the categories for the products of these classes were directly under the "Official Hardware" top level
category. They are hereby moved under the appropriate "Mega" product class parent category.