Add "Pro Solutions" product class category
The subcategories of the "Official Hardware" category are to be organized under a subcategory for each of the product
classes. These product classes are defined by the hardware documentation home page.

That hardware documentation places some products under a class named "Pro Solutions", so a forum category of this name
is added.