Deprecate "UNO R4" category
The subcategories of the "Official Hardware" category are to be organized under a subcategory for each of the product
classes. These product classes are defined by the hardware documentation home page.

That hardware documentation places the UNO R4 boards under products under a class named "Classic".

The child categories for each of the boards have already been moved to the appropriate "Classic" parent category. The
category was purely a container, it no longer contains any posts.

This is an interim measure until we can set up a redirect from the "UNO R4" category URL to the "Classic" category in
order to avoid breaking any links. Once that is done, the "UNO R4" category will be deleted.