Add "Retired" product class category
The subcategories of the "Official Hardware" category are to be organized under a subcategory for each of the product
classes.

However, this classification is not done for hardware products which are no longer manufactured. A "Retired" category
will be used as the container for the existing categories for retired products.