Use standardized name for Arduino Education products category
The subcategories of the "Official Hardware" category are to be organized under a subcategory for each of the product
classes. These product classes are defined by the hardware documentation home page.

That hardware documentation places the Arduino Education products under a class named "Education", so the parent
category for those products is renamed accordingly.