Deprecate "Arduino SIM" category
The Arduino SIM service was discontinued some time ago, which means that no new topics will be created on this subject.
A survey of the contents of the category revealed that most of the topics were only about cellular communication in
general, created by those who didn't even use or have interest in the Arduino SIM. Since Arduino SIM was exclusively a
connectivity solution for Arduino Cloud IoT Devices, which could not be used for general communication, the "IoT Cloud"
category is suitable for the on-topic topics in the category.

So the category has been deemed superfluous. All posts have been moved to the appropriate alternative category.

This is an interim measure until we can set up a redirect from the "Arduino SIM" category URL to the "IoT Cloud"
category in order to avoid breaking any links. Once that is done, the "Arduino SIM" category will be deleted.