Deprecate "Intel®-based platforms" category
A survey of the contents of the category revealed that there was very little usage, and many of the topics were actually about irrelevant subjects. In addition, although Arduino Cloud still supports the use of these platforms as IoT Devices, the nature of that support has changed greatly, rendering the context of the category outdated.

So the category has been deemed superfluous. All topics have been moved to the appropriate alternative category.

This is an interim measure until we can set up a redirect from the "Intel®-based platforms" category URL to the "IoT
Cloud" category in order to avoid breaking any links. Once that is done, the "Intel®-based platforms" category will be deleted.