Deprecate "Emergency Response: Covid-19 Projects" category
A survey of the contents of the category revealed that there was very little ongoing usage, and many of the recent
topics were actually about irrelevant subjects.

So the category has been deemed superfluous. All topics have been moved to the appropriate alternative category.

This is an interim measure until we can set up a redirect from the "Emergency Response: Covid-19 Projects" category URL
to the "Showcase" category in order to avoid breaking any links. Once that is done, the "Emergency Response: Covid-19
Projects" category will be deleted.