Close deprecated categories
Several categories have been deemed superfluous, and will be deleted.

Some tasks need to be completed before that can be done:

- Move all topics to the remaining alternative categories
- Set up a redirect from the category URL to an equivalent page

As it will be some time before these tasks are completed, the categories are being configured to prevent the creation of
new topics (which would only have to be moved by a forum maintainer). In the categories that still contain topics, the
permission to reply to existing topics is preserved in order to avoid interference with ongoing discussions.