Use more descriptive name for "Chrome App" category
Arduino created several Chrome apps to support the use of Arduino Cloud from ChromeOS machines. The "Chrome App"
category was created to support the users of those apps.

The chosen category name is too ambiguous. The convention has been established of dropping the "Arduino" branding from
product titles in forum categories to avoid unnecessary verbosity, but this would imply the product name was "Arduino
Chrome App", which it is not.

The new name "Cloud Chrome App" is not entirely accurate since the apps are named "Arduino Create" and "Arduino Create
for Education". However, the service has been renamed from "Arduino Create" to "Arduino Cloud", and so in general users
will not be familiar with the "Create" term. Support was added for using the browser-based Cloud Editor from ChromeOS
machines. The Chrome apps have been deprecated and now apparently only open Cloud Editor in the browser when used. So
the category is likely only of interest for historical purposes and the "Cloud Editor" category will be more appropriate
for any new topics from ChromeOS users. For this reason, it is more important for the category name to communicate the
subject matter of the category to forum users than for it to exactly match the obsolete product name.