Correct category asset folder names
The folders that contain the category assets are named according to the "slug" assigned to the category in the forum
configuration, which is used to make forum URLs more human (and SEO?) friendly.

Some of the folder names were incorrect or inaccurate, either due to a mistake during the creation of the folder in this
repository, or a mistake during the creation of the creation of the category on the forum, which has since been
corrected.