Standardize category description topic titles
Each forum category contains a special topic which serves two purposes:

- Define the description that is shown along with the category name in various parts of the forum UI
- Document the purpose of the topic

The forum software gives these topics a title with the format "About the ___ category", and that has been adopted as the
standard for Arduino Forum. Such standardization is important in order to allow the forum users to easily find the
information about the purpose of each category so that they can determine whether it is appropriate for their topic.

Previously, the titles of some of these topics had been changed to have an "About the ___" format. In addition to being
inconsistent, this also did not correctly reflect the true purpose of the topic.