Deprecate "Cloud" category
This category was created during the early days of the Arduino Cloud project. It was later superseded by the "Cloud
Editor" and "Cloud IoT" categories. So there is no reason for this category to continue to exist.

 All topics have been moved to the appropriate alternative category.

This is an interim measure until we can set up a redirect from the "Cloud" category URL to the "Cloud IoT" category in
order to avoid breaking any links. Once that is done, the "Cloud" category will be deleted.