Add "Moved to Nederlands category" post template
A common forum maintenance task is recategorization of topics written in languages for which a dedicated category
exists, but published under the English language categories.

It is useful for the maintainer to add a reply to notify the user of the recategorization and request they select an
appropriate category in the future. Ideally, this reply should be written in the appropriate language.

Forum user sterretje has contributed this Dutch language version of the reply.