Adjust placeholder location to gracefully handle non-replacement
The "Moved to Nederlands category" post template contains a placeholder which the template user should replace with a username mention at the time of posting.

The template user might forget to make this replacement. The placeholder is in the form of an HTML comment tag, so will simply not be rendered in the published post if this occurs. With the previous placement of the placeholder at the start of a sentence, the published post would have an obviously incorrect form. By moving the placeholder to the end of the sentence, this eventually is mitigated down to only a single excess space in the published post.

The new sentence form is grammatically valid and approximately equivalent to the previous form.