Deprecate "Windows Remote Arduino" category
A survey of the contents of the category revealed that there was very little usage, and many of the topics were actually
about irrelevant subjects. The "Windows Remote Arduino" project was abandoned by Microsoft years ago. Arduino carried on
the spirit of this work in what is now the Arduino Cloud "Manager for Linux", but the "Windows Remote Arduino" category
is of little relevance to that service and the existing "Cloud IoT" category is sufficient for topics on that subject.

So the category has been deemed superfluous. All topics have been moved to the appropriate alternative category.

This is an interim measure until we can set up a redirect from the "Windows Remote Arduino" category URL to the
"Interfacing w/ Software on the Computer" category in order to avoid breaking any links. Once that is done, the "Windows
Remote Arduino" category will be deleted.