Deprecate "Windows Virtual Shields for Arduino" category
A survey of the contents of the category revealed that there was very little usage, and many of the topics were actually
about irrelevant subjects. The "Windows Virtual Shields for Arduino" project was abandoned by Microsoft years ago.

So the category has been deemed superfluous. All topics have been moved to the appropriate alternative category.

This is an interim measure until we can set up a redirect from the "Windows Virtual Shields for Arduino" category URL to
the "Interfacing w/ Software on the Computer" category in order to avoid breaking any links. Once that is done, the
"Windows Virtual Shields for Arduino" category will be deleted.