Move "Staff" category to lower position in category list
The "Staff" category is used for internal discussions by Arduino Forum staff (moderators and administrators). Although
the category does serve an important purpose, it is infrequently used.

Previously the category was placed at the very top of the category list. This prominent placement was not proportionate
to its usage. By moving it to a position below the categories that are more frequently used by staff members, the user
experience will be improved for the staff members.