Restore visibility of deprecated categories
Previously, in order to reduce the clutter of the forum's category structure, the permissions of these deprecated
categories were configured to cause them to not be publicly visible. Since then, the alternative approach has been
adopted of moving deprecated categories under a "Recycle Bin" category, which is public but not shown in the primary
topic listings. The previous approach is potentially harmful in that it would break any links to the category itself
(though not to the posts that had been moved out of the category, since post URLs are not category-dependent). Making
the categories visible once more avoids the breakage of any links to the category.