Remove obsolete entries from link check ignore configuration
The project infrastructure includes a check for broken links. Some of the links in the content are to pages which are
not publicly accessible. These would cause spurious failures of the link check, so the link checker tool must be
configured to ignore those URLs.

Some of the ignore entries are no longer necessary, either due to the page having been made publicly accessible, or the
link removed from the content. These obsolete entries are hereby removed from the link checker configuration to keep it
tidy and ensure comprehensive coverage of the links.