Add Dutch language user guide link to "Moved to Nederlands category" template
The "Moved to Nederlands category" post template contains a link to the forum user guide.

Previously, due to the Dutch language guide being outdated, the template only linked to the English language guide. A
modern Dutch language guide has now been published, and so a link to that guide is now added to the template.

---

Translation contributed by forum user sterretje.