Correct name of "Русский" category
The majority of the forum activity is conducted in the English language. However, some forum users prefer to use a
different language. Dedicated categories have been created for several non-English languages.

Previously, the category for the Russian language was incorrectly named "Россия - Russia" ("Россия" translates to
"Russia"). This name was incorrect because these categories are about language, not geographic location. The previous
category name gave the impression that the category was dedicated to topic related to the subject of the country of
Russia, and did not clearly communicate the true purpose of being a place for forum users to communicate in Russian.
Multiple countries other than Russia use the Russian language, and the language is also used by people living in
countries where Russian is not the primary language.

The category name is hereby corrected to "Русский" (translation: "Russian"), which accurately communicates the purpose
of the category. The "- Russia" suffix was removed entirely as this does not align with the common convention for the
non-English language category names.