Deprecate "Interactive Art" category
A survey of the contents of the category revealed that there was very little usage, and many of the topics were actually
about irrelevant subjects or subjects for which a more appropriate category exists.

So the category has been deemed superfluous. All topics have been moved to the appropriate alternative category.

This is an interim measure until we can set up a redirect from the "Interactive Art" category URL to the "Showcase"
category in order to avoid breaking any links. Once that is done, the "Interactive Art" category will be deleted.