Correct microcontroller name in "About the Nano ESP32 category" post
The Nano ESP32 board uses the ESP32-S3 microcontroller. The post previously stated that it uses "the ESP32"
microcontroller. Although the ESP32-S3 is one model in the ESP32 family of microcontrollers, it is not accurate to say
the board uses "the ESP32" microcontroller, as that implies it uses the classic ESP32 (which is another specific model,
which has the same name as the family).