Update formatting to Prettier 3.6.2 style
Versions of Prettier <3.6.0 added an unnecessary empty line in Markdown nested lists under certain conditions. That bug
was fixed in version 3.6.0, and the content updated to match the resulting formatting at the time of the bump to that
version.

However, it seems that some of empty lines removed by Prettier 3.6.0 were actually wanted after all (?), as the behavior
of Prettier was changed once again in version 3.6.2, adding back some (but not all) of the empty lines that were removed
by 3.6.0.