Move topic split instructions to dedicated procedure in "Moderator Instructions"
Previously, the topic split instructions were provided as part of the "Split hijacked topic" procedure.

There are cases other than a hijack where splitting a topic is useful, so it is better to define a general purpose
"Split topic" procedure, then reference that from each of the more specific procedures that require a topic split.