Update slug of "Arduino Command Line Tools" category
This category was originally named "Arduino CLI", and limited to the subject of that specific tool. The category's scope
was later expanded to all official command line tools, but the category slug was not updated at that time.

Discourse automatically redirects category URLs (the only relevant part is the ID path segment, and the ID is not
affected by this change), so this will not break any existing links to URLs containing the previous slug.