Update "About" post to reflect true scope of "Arduino Command Line Tools" category
This category was originally named "Arduino CLI", and limited to the subject of that specific tool. The category's scope
was later expanded to all official command line tools, but the category's "About the Arduino Command Line Tools
category" post was not updated at that time.

The post is hereby updated to accurately describe the purpose of the category.