Rename "Accessories" category to "Modulino"
The subcategories of the "Official Hardware" category use the structure established by the documentation website.

Previously, the categories for the Modulino products were placed under an "Accessories" category. This was inconsistent
with the documentation website, which places them under a top level "Modulino" category.

Since the "Accessories" category only contained Modulino product categories, and since there is no "Accessories"
category on the documentation website, it can be reused for this purpose.