Merge individual Modulino product categories
Previously, a dedicated category was created for each of the Arduino Modulino products. This is an established approach
which generally works well. However, over the course of >6 months of the existence of these categories, they have only
been used a few times (with some categories going completely unused).

A turning point has been reached with the release of four new Modulino products. Previously when there were only seven
products it was reasonable to have dedicated categories even if rarely utilized. However, having a large number of
categories does have a harmful effect on the usability of the forum. As the need arose to increase that number to eleven
(and quite possibly beyond that in the future), it seems likely the harm caused by these categories outweighs the
benefits. For this reason, the decision was made to use a single category for all Modulino products.

The user created topics were moved from the per-product categories to the general "Modulino" category.

In order to avoid breaking any links to the existing categories, they were moved under a "recycle bin" category
dedicated to deprecated categories. The hope is to eventually remove these categories if a way is found to create
redirects to maintain the functionality of links targeting their URLs. The category descriptions are updated to clearly
communicate this situation and direct users to the "Modulino" category.